CodeYard
A full-stack Web Blog based on the flask microframework, using PHPMyAdmin as a back-end Database. The Admin Login is 2-Step, using sessions OTPs for verification. Queries when submitted trigger a mail that acknowledges the reception of the query. Kindly feel free to use the template, just ensure that you edit the Config.json to your needs. It also features a file uploader that secures the filename before it is actually uploaded. Admin Panel allows user to Add Posts, Delete Posts, and add files.